Add process & tools for building Droplet image
The OpenVPN + Pi-Hole marketplace image [0] is fairly popular. I built
this a few years ago on Bionic for a hackathon. However, a recent
issue [1] identified that Pi-Hole no longer works on Bionic.

Rather than invest time to update this image, I've taken the opportunity
to build a new image that incorporates many of the feature requests over
the years, and some nice to haves.

- IPV6: https://github.com/digitalocean/OpenVPN-Pihole/issues/2
- No Expiration: https://github.com/digitalocean/OpenVPN-Pihole/issues/7
- Multi-ports: https://github.com/digitalocean/OpenVPN-Pihole/issues/8
- $4 Droplet: https://github.com/digitalocean/OpenVPN-Pihole/issues/33
- WireGuard: WireGuard is a faster, easier to setup and configure VPN
- Easier Multi-Client: QR code scanning for quick and easy client add
- Unbound: Local recursive DNS on the server

[0] https://github.com/digitalocean/OpenVPN-Pihole
[1] https://github.com/digitalocean/OpenVPN-Pihole/issues/37